requirements.txt
dnspython
pike
requests
PyYAML
six
